rating draft

+++ src/main/java/dev/lukaszmichalak/importer/DbCache.java
@@ -2,4 +2,4 @@ package dev.lukaszmichalak.importer;
 
 import java.util.Map;
 
-record DbCache(Map<String, Integer> voivodeshipMap, Map<String, Integer> productTypeMap) {}
+record DbCache(Map<String, Long> voivodeshipMap, Map<String, Long> productTypeMap) {}
