+++ src/main/java/dev/lukaszmichalak/regionalproducts/product/ProductRepository.java
@@ -4,11 +4,11 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.repository.ListCrudRepository;
 
-interface ProductRepository extends ListCrudRepository<Product, Integer> {
+interface ProductRepository extends ListCrudRepository<Product, Long> {
 
   Optional<Product> findByName(String name);
 
-  List<Product> findByVoivodeshipId(Integer name);
+  List<Product> findByVoivodeshipId(Long name);
 
-  long countByVoivodeshipId(Integer id);
+  long countByVoivodeshipId(Long id);
 }
