+++ src/main/java/dev/lukaszmichalak/regionalproducts/product/exception/ProductNotFoundException.java
@@ -2,7 +2,7 @@ package dev.lukaszmichalak.regionalproducts.product.exception;
 
 public class ProductNotFoundException extends RuntimeException {
 
-  public ProductNotFoundException(Integer id) {
+  public ProductNotFoundException(Long id) {
     super("Nie znaleziono produktu o id: " + id);
   }
