+++ src/main/java/dev/lukaszmichalak/regionalproducts/statistics/StatisticsRepository.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 import java.util.Optional;
 import org.springframework.data.repository.ListCrudRepository;
 
-interface StatisticsRepository extends ListCrudRepository<Statistics, Integer> {
+interface StatisticsRepository extends ListCrudRepository<Statistics, Long> {
 
-  Optional<Statistics> findByVoivodeshipIdAndDate(Integer voivodeshipId, LocalDate date);
+  Optional<Statistics> findByVoivodeshipIdAndDate(Long voivodeshipId, LocalDate date);
 }
