+++ src/main/java/dev/lukaszmichalak/regionalproducts/voivodeship/VoivodeshipRepository.java
@@ -3,7 +3,7 @@ package dev.lukaszmichalak.regionalproducts.voivodeship;
 import java.util.Optional;
 import org.springframework.data.repository.ListCrudRepository;
 
-interface VoivodeshipRepository extends ListCrudRepository<Voivodeship, Integer> {
+interface VoivodeshipRepository extends ListCrudRepository<Voivodeship, Long> {
 
   Optional<Voivodeship> findByCodeIgnoreCase(String code);
 }
